feat: implement SyncStateService and update SyncSession schema

+++ MarketplaceSyncer.Service/BusinessRu/Models/Responses/GroupResponse.cs
@@ -3,8 +3,8 @@ using System.Text.Json.Serialization;
 namespace MarketplaceSyncer.Service.BusinessRu.Models.Responses;
 
 public record GroupResponse(
-    [property: JsonPropertyName("id")] int Id,
+    [property: JsonPropertyName("id")] long Id,
     [property: JsonPropertyName("name")] string Name,
-    [property: JsonPropertyName("parent_id")] int? ParentId,
+    [property: JsonPropertyName("parent_id")] long? ParentId,
     [property: JsonPropertyName("archive")] bool IsArchive
 );
